test(cli): stabilize cli tests
